refactor : modify book view model file name
