+++ lib/presentation/book/seat/seat_screen.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:griffin/presentation/book/seat/seat_view_model.dart';
@@ -6,16 +7,18 @@ import 'package:griffin/presentation/book/seat/widgets/column_number_widget.dart
 import 'package:griffin/presentation/book/seat/widgets/seat_container_widget.dart';
 import 'package:griffin/presentation/book/seat/widgets/seat_label_widget.dart';
 import 'package:provider/provider.dart';
+
 import '../../../domain/model/books/books_model.dart';
 import '../../common/colors.dart';
 import '../../common/common_button.dart';
 
 class SeatScreen extends StatefulWidget {
   final List<BooksModel> bookIdList;
-  final int totalFare;
 
-  const SeatScreen(
-      {super.key, required this.bookIdList, required this.totalFare});
+  const SeatScreen({
+    super.key,
+    required this.bookIdList,
+  });
 
   @override
   State<SeatScreen> createState() => _SeatScreenState();
@@ -31,7 +34,7 @@ class _SeatScreenState extends State<SeatScreen> with TickerProviderStateMixin {
     super.initState();
     Future.microtask(() {
       final seatViewModel = context.read<SeatViewModel>();
-      seatViewModel.init(widget.totalFare);
+      seatViewModel.init();
     });
     _tabController = TabController(
       length: 2,
@@ -125,7 +128,8 @@ class _SeatScreenState extends State<SeatScreen> with TickerProviderStateMixin {
                                 selectedSeatList.length < _numberOfPeople,
                             index: index,
                             list: selectedSeatList,
-                            bookIdListLength: widget.bookIdList.length, isDeparture:  _tabController.index == 0,
+                            bookIdListLength: widget.bookIdList.length,
+                            isDeparture: _tabController.index == 0,
                           );
                         } else if (adjustedIndex < 21) {
                           // return BusinessClass();
@@ -135,7 +139,8 @@ class _SeatScreenState extends State<SeatScreen> with TickerProviderStateMixin {
                                 selectedSeatList.length < _numberOfPeople,
                             list: selectedSeatList,
                             index: index,
-                            bookIdListLength: widget.bookIdList.length, isDeparture:  _tabController.index == 0,
+                            bookIdListLength: widget.bookIdList.length,
+                            isDeparture: _tabController.index == 0,
                           );
                         } else {
                           // return EconomyClass();
@@ -145,7 +150,8 @@ class _SeatScreenState extends State<SeatScreen> with TickerProviderStateMixin {
                                 selectedSeatList.length < _numberOfPeople,
                             list: selectedSeatList,
                             index: index,
-                            bookIdListLength: widget.bookIdList.length, isDeparture:  _tabController.index == 0,
+                            bookIdListLength: widget.bookIdList.length,
+                            isDeparture: _tabController.index == 0,
                           );
                         }
                       }),
@@ -189,23 +195,27 @@ class _SeatScreenState extends State<SeatScreen> with TickerProviderStateMixin {
                                   style: TextStyle(fontSize: 20),
                                 ),
                                 const SizedBox(height: 6),
-                               Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                                    children: [
-                                      TextButton(
-                                        onPressed: () {},
-                                        child: const Text('NO'),
-                                      ),
-                                      TextButton(
-                                        onPressed: ()  {
-                                          if (mounted) {
-                                          context.push('/book/passport/seat', extra: {"bookIdList": widget.bookIdList});
+                                Row(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceEvenly,
+                                  children: [
+                                    TextButton(
+                                      onPressed: () {},
+                                      child: const Text('NO'),
+                                    ),
+                                    TextButton(
+                                      onPressed: () {
+                                        if (mounted) {
+                                          context.push('/book/passport/seat',
+                                              extra: {
+                                                "bookIdList": widget.bookIdList
+                                              });
                                         }
-                                          },
-                                        child: const Text('YES'),
-                                      ),
-                                    ],
-                                  ),
+                                      },
+                                      child: const Text('YES'),
+                                    ),
+                                  ],
+                                ),
                               ],
                             ),
                           ),
