+++ lib/presentation/book/seat/seat_view_model.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:griffin/presentation/book/seat/seat_state.dart';
+
 import '../../../data/core/result.dart';
 import '../../../domain/model/books/books_model.dart';
 import '../../../domain/model/user/user_account_model.dart';
@@ -23,8 +25,8 @@ class SeatViewModel extends ChangeNotifier {
 
   SeatState get state => _state;
 
-  void init(int totalFare) async {
-    _state = _state.copyWith(totalFare: totalFare);
+  void init() async {
+    // _state = _state.copyWith(totalFare: totalFare);
     await getSession();
   }
